Benchmark REPS inherit from optimizer REPS
